+++ CHANGES.rst
@@ -1,3 +1,11 @@
+0.0.3
+=====
+
+- Add atmospheric corrections.
+- Add simple Bouguer and spherical Bouguer corrections.
+- Add gravity field of a prism.
+- Add `coordinates.position.Position3D` class.
+
 0.0.2
 =====